ffmpegManager に ffmpeg の処理を分割
